Add support for Hubot Prefxes
This commit adds support for Hubot to have an alias associated with
its usage. This allows for easier typing of commands in chat without
having to address the full name of the bot for each command. By
default this alias will be set to '/', and can be changed by
setting the `hubot::chat_alias` attribute in Hiera

This commit changes the version of Hubot module to a fork until
https://github.com/evenup/evenup-hubot/pull/18 is merged upstream