update regex to generate environments
This commit updates the helper function that generates
environments based on git branches to include branches
that have dots (.) in the branch name.

This also ensures that at the CLI a user does not have
to transpose the ENV to a valid puppet environment, either.

ENV=migrate-mistral-0.14dev will be converted to
ENV=migrate_mistral_0_14dev